added fixed power mismatch code (needs testing)
It's not an easy bit of code but I have copied the core routine from Laos where it was tested. The problem might lie in setting up the variables to use it. It needs a manual test of setting the power to 50% and again to something that isn't exactly half.